Cast the timeout when fetching from JetStream
We control our timeout via environment variables and had a silly bug in our code where the timeout was sent as a string. We kept running out of memory because the expires line was copying our string ("5") one billion times.

This is just a simple fix to save new users some hassle.